requirements.txt
accelerate>=0.16.0
argparse
csv
datasets>=2.19.1
datasets[vision]
diffusers
distutils
ftfy
googleapiclient.discovery
huggingface_hub
io
Jinja2
numpy
os
packaging
pandas
pathlib
peft==0.7.0
PIL
pillow
re
requests
setuptools
tensorboard
torch
torchvision
transformers>=4.25.1
torch
tqdm
